2186. Minimum Number of Steps to Make Two Strings Anagram II
