+++ src/test/kotlin/me/bossm0n5t3r/leetcode/minimumNumberOfStepsToMakeTwoStringsAnagram2/MinimumNumberOfStepsToMakeTwoStringsAnagram2Test.kt
@@ -1,10 +1,10 @@
-package minimumNumberOfStepsToMakeTwoStringsAnagram2
+package me.bossm0n5t3r.leetcode.minimumNumberOfStepsToMakeTwoStringsAnagram2
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 internal class MinimumNumberOfStepsToMakeTwoStringsAnagram2Test {
-    private val minimumNumberOfStepsToMakeTwoStringsAnagram2 = MinimumNumberOfStepsToMakeTwoStringsAnagram2()
+    private val minimumNumberOfStepsToMakeTwoStringsAnagram2 = MinimumNumberOfStepsToMakeTwoStringsAnagram2.Solution()
 
     private data class MinimumNumberOfStepsToMakeTwoStringsAnagram2TestData(
         val s: String,
